requirements.txt
Flask==3.0.2
Flask-CORS==4.0.0
bcrypt==4.1.2
Werkzeug==3.0.1
python-frontmatter==1.0.0
markdown==3.5.1
deep-translator==1.11.4
PyYAML==6.0.1
python-i18n==0.3.9
gunicorn==21.2.0
beautifulsoup4==4.13.0
requests==2.31.0
pymongo[srv]>=4.6.1
python-magic==0.4.27
PyJWT==2.8.0